+++ MST13/Assets/Effect.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 33f027e4222ebbe4395f6aa5354e6bb9
+folderAsset: yes
+timeCreated: 1510550753
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
