+++ MST13/Assets/Effect/Slash.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4e3d75a0ed7df854ba3165d89eec8a76
+folderAsset: yes
+timeCreated: 1510550765
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
