+++ MST13/Assets/Effect/Slash/SlashMaterial.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f2ea9967fc724774596682b6a43992ac
+timeCreated: 1510550811
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
