+++ MST13/Assets/Effect/Slash/TrailSlashMaterial.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 1c973533a4c1de64e93c30020ac0a053
+timeCreated: 1510552057
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
