+++ MST13/Assets/Effect/Slash/backSlashMaterial.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4c65f736358195a4580833a0f97f40ba
+timeCreated: 1510554052
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
