+++ MST13/Assets/Effect/Sword.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f855f1f3943507d49b0a8f9aae65066d
+timeCreated: 1510644748
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: -1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
