+++ MST13/Assets/MSTAsseats/Animation.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 1ca797746e3ef2847a76c565713c65ff
+folderAsset: yes
+timeCreated: 1511254609
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
