+++ MST13/Assets/MSTAsseats/Animation/RotEffect.anim.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7eba4baa3f0f8324796e38b505e3d577
+timeCreated: 1511254609
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
