+++ MST13/Assets/MSTAsseats/Animation/bottomBlast.controller.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 087b2092aa63c5f48a4fae596f883650
+timeCreated: 1511254611
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
