+++ MST13/Assets/MSTAsseats/Scene/ParticleEffect.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcc9582adac81b340a1ecdb7873d0b86
+timeCreated: 1510644306
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
