+++ MST13/Assets/MSTAsseats/Scene/TestTime.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 55d1f138c5e56384bb65bf4a4c3253bc
+folderAsset: yes
+timeCreated: 1509600795
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
